Update reports functionality (store, update, destroy)

+++ app/Http/Controllers/ReportController.php
@@ -21,8 +21,6 @@ class ReportController extends Controller
     $lostModal = Auth::check() ? '#createLostAnimalReportModal' : '#loginReminderModal';
     $foundModal = Auth::check() ? '#createFoundAnimalReportModal' : '#loginReminderModal';
 
-    // return view('reports.index', compact('reports','user','lostModal','foundModal'));
-
     return Inertia::render('Reports/Index',[
       'lostModal' => $lostModal,
       'foundModal' => $foundModal,
@@ -54,7 +52,7 @@ class ReportController extends Controller
 
     $report = Report::create($requestData);
 
-    return redirect()->back()->with('success', $report->getTypeFormattedAttribute() . ' Report created successfully!');
+    return redirect()->back()->with('success', $report->getTypeFormattedAttribute() . ' Report has been created!');
   }
 
   /**
@@ -81,15 +79,18 @@ class ReportController extends Controller
     $requestData = $request->all();
 
     if ($request->hasFile('image')) {
-      Storage::delete($report->image);
-
+      if($report->image){
+        Storage::delete($report->image);
+      }
       $imagePath = $request->file('image')->store('images/reports/reports_images', 'public');
-
       $requestData['image'] = $imagePath;
+    }else{
+      unset($requestData['image']);
     }
 
     $report->update($requestData);
-    return redirect()->route('reports.index')->with('success', $report->getTypeFormattedAttribute() . ' Report updated successfully!');
+    
+    return redirect()->route('reports.index')->with('info', $report->getTypeFormattedAttribute() . ' Report updated successfully!');
   }
 
   /**
@@ -99,6 +100,6 @@ class ReportController extends Controller
   {
     $report->delete();
 
-    return redirect()->route('reports.index')->with('success', $report->getTypeFormattedAttribute() . ' Report has been deleted successfully!');
+    return redirect()->route('reports.index')->with('warning', $report->getTypeFormattedAttribute() . ' Report has been deleted successfully!');
   }
 }
